services/parser-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
rq==1.16.2
redis==5.0.4
pymupdf==1.25.3
python-docx==1.1.2
boto3==1.34.89
psycopg2-binary==2.9.10
prometheus-fastapi-instrumentator==6.1.0
pydantic-settings==2.2.1
